Beverage 관련 프로토콜 / 데이터타입을 별개의 파일로 분리
