+++ VendingMachineApp/VendingMachineApp/Beverage.swift
@@ -27,39 +27,3 @@ class Beverage: CustomStringConvertible {
         return "[\(brand)] \(name): ₩\(price), \(size)ml (\(manufactured.inFormat("yyMMdd")))"
     }
 }
-
-enum Container {
-    case can
-    case bottle
-}
-
-protocol Transportable {
-    
-    func containerType() -> Container
-    
-    func isTransportable() -> Bool
-    
-}
-
-protocol Expirable {
-    
-    func isExpired() -> Bool
-    
-}
-
-protocol LactoFreeable {
-    
-    func isLactoFree() -> Bool 
-}
-
-protocol SugarFreeable {
-    
-    func isSugarFree() -> Bool
-    
-}
-
-protocol Hotable {
-    
-    func isHotable() -> Bool
-    
-}
