+++ VendingMachineApp/VendingMachineApp/Package.swift
@@ -1,0 +1,14 @@
+//
+//  Package.swift
+//  VendingMachineApp
+//
+//  Created by Song on 2021/02/25.
+//
+
+import Foundation
+
+enum Package {
+    case can
+    case bottle
+    case carton
+}
